# public/openapi.yaml
openapi: 3.0.1
info:
  title: Kaguya
  description: A project that you can edit just by talking to ChatGPT.
  version: v1
servers:
  - url: http://localhost:3000
paths:
  /api/executeCommand:
    post:
      operationId: executeCommand
      summary: Execute a shell command - from git commands to running code, listing files, or anything else that's possible to do through a shell command.
      description: This endpoint allows users to execute arbitrary shell commands. Note, 'git push' is not available through this endpoint due to authentication requirements.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                command:
                  type: string
                  description: The shell command to execute.
              required:
                - command
      responses:
        "200":
          description: Command executed successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating the success of the command execution.
                  output:
                    type: string
                    description: The output of the executed command.
        "400":
          description: Bad request (e.g., missing command parameter).
        "500":
          description: Internal server error (e.g., error executing command).
  /api/listFilesInDirectory:
    get:
      parameters:
        - in: query
          name: directoryPath
          schema:
            type: string
          description: The path of the directory to list files and directories in.
      operationId: listFilesInDirectory
      summary: List files and directories in the specified directory. Files and directories specified in .gitignore files will be ignored. To specify the root directory, use ".".
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: string
                    description: An array of file paths
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: An error message
  /api/readFile:
    get:
      operationId: readFile
      summary: Read the content of a file in the user's directory.
      parameters:
        - in: query
          name: filePath
          schema:
            type: string
          required: true
          description: The path of the file to read
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
                    description: The content of the file
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: An error message
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: An error message
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: An error message
...